Fix a potential crash
The issue is that the module closes steamapi.so lib handle (value passed in SteamAPI_Init_internal hook), where it should not.
When steamapi.so is not loaded from the server root directory (e.g. /.steam/sdk32), a crash happens all the time.
Not totally sure why it doesn't crash if when the file is loaded from server root directory.